requirements.txt
asgiref==3.3.4
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.23
dj-database-url==0.5.0
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
ephem==3.7.7.1
gunicorn==20.1.0
hijri-converter==2.1.2
holidays==0.11.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pkg-resources==0.0.0
plotly==4.14.3
prophet==1.0.1
psycopg2==2.8.6
PyMeeus==0.5.11
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
ujson==4.0.2
